The check for the cleanup flag in $1 also matches --full-static
There is an ambiguity in how short and long cli flags are passed that
causes the script to clean whenever a full-static build is attempted
with long cli option `--full-static``.

The second clause of the conditional that checks for the cleanup flag

    if [[ "$1" == "--cleanup" || "$1" =~ 'c' ]]; then

is overly broad and evaluates as truthy if $1 is "--full-cleanup". This
can be easily fixed by appending an additional string check against the
string "==", e.g `! "$1" =~ '--'`.